refactor kebab, remove -default suffix
