+++ packages/theme/scripts/build-json.mjs
@@ -1,6 +1,6 @@
-import { kebabCase } from "change-case";
 import { StyleDictionary } from "style-dictionary-utils";
 import { cssMultiModes } from "./sd-utils/format/css-multi-modes.mjs";
+import { saltKebab } from "./sd-utils/transform/kebab.mjs";
 
 // TODO: add more styling options here, and theme vs theme next
 function getStyleDictionaryConfig({ modes, density, accents }) {
@@ -142,16 +142,6 @@ function getStyleDictionaryConfig({ modes, density, accents }) {
   };
 }
 
-// Regexps involved with splitting words in various case formats.
-const SPLIT_LOWER_UPPER_RE = /([\p{Ll}\d])(\p{Lu})/gu; // ( lower case + digit ) ( upper case )
-const SPLIT_LOWER_NON_DIGIT_UPPER_RE = /([\p{Ll}])(\p{Lu})/gu; // ( lower case  ) ( upper case )
-const SPLIT_UPPER_UPPER_RE = /(\p{Lu})([\p{Lu}][\p{Ll}])/gu; // ( upper case ) ( [ upper case ] [ lower case ] )
-// The replacement value for splits.
-const SPLIT_REPLACE_VALUE = "$1\0$2";
-
-// Regexp involved with stripping non-word characters from the result.
-const DEFAULT_STRIP_REGEXP = /[^\p{L}\d]+/giu;
-
 const modes = ["light", "dark"];
 const accents = ["blue", "teal"];
 // for await (const mode of ["dark"]) {
@@ -206,73 +196,7 @@ for await (const density of ["hd"]) {
     name: "salt-ds/name/kebab",
     type: "name",
     transitive: true,
-    transform: (token, config) => {
-      // attributes: { category: 'color', type: 'alpha',
-      if (
-        token.path.includes("alpha") &&
-        token.attributes.category === "color" &&
-        token.attributes.type === "alpha"
-      ) {
-        const alphaIndex = token.path.findIndex((p) => p === "alpha");
-        const alphaRemovedPath = [
-          ...token.path.slice(0, alphaIndex),
-          ...token.path.slice(alphaIndex + 1),
-        ];
-        // console.log("Token path with alpha removed", alphaRemovedPath);
-
-        function splitPrefixSuffix(input) {
-          const splitFn = modifiedSplit;
-          const prefixIndex = 0;
-          const suffixIndex = input.length;
-
-          return [
-            input.slice(0, prefixIndex),
-            splitFn(input.slice(prefixIndex, suffixIndex)),
-            input.slice(suffixIndex),
-          ];
-        }
-
-        function specialKebab(input) {
-          const [prefix, words, suffix] = splitPrefixSuffix(input);
-          return (
-            prefix +
-            words.map((input) => input.toLowerCase()).join("-") +
-            suffix
-          );
-        }
-
-        function modifiedSplit(value) {
-          let result = value.trim();
-
-          result = result
-            // `SPLIT_LOWER_NON_DIGIT_UPPER_RE` changed compare with 'change-case' original split
-            // Change to not split 30A -> 30-A
-            .replace(SPLIT_LOWER_NON_DIGIT_UPPER_RE, SPLIT_REPLACE_VALUE)
-            .replace(SPLIT_UPPER_UPPER_RE, SPLIT_REPLACE_VALUE);
-
-          result = result.replace(DEFAULT_STRIP_REGEXP, "\0");
-
-          let start = 0;
-          let end = result.length;
-
-          // Trim the delimiter from around the output string.
-          while (result.charAt(start) === "\0") start++;
-          if (start === end) return [];
-          while (result.charAt(end - 1) === "\0") end--;
-
-          return result.slice(start, end).split(/\0/g);
-        }
-
-        const name = specialKebab(
-          [config.prefix].concat(alphaRemovedPath).join(" "),
-        );
-        console.log("salt-ds/name/kebab specialKebab name", name, token);
-        return name;
-      }
-      const name = kebabCase([config.prefix].concat(token.path).join(" "));
-      console.log("salt-ds/name/kebab name", name, token);
-      return name;
-    },
+    transform: saltKebab,
   });
   const sd = await saltStyleDictionary.extend(config);
   sd.cleanAllPlatforms();
